Namespace Refactor, 2nd pass. Updated to .NET Framework 4.8 for additional C# Language features and optimizations.
