+++ FalconBMS Alternative Launcher Cs/App.xaml.cs
@@ -3,6 +3,8 @@ using System.Diagnostics;
 using System.Windows;
 using System.Windows.Threading;
 
+using FalconBMS.Launcher.Core;
+
 namespace FalconBMS.Launcher
 {
     /// <summary>
