+++ FalconBMS Alternative Launcher Cs/Core/TheaterList.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Windows.Controls;
 
-namespace FalconBMS.Launcher
+namespace FalconBMS.Launcher.Core
 {
     public class TheaterList
     {
