+++ FalconBMS Alternative Launcher Cs/Input/MainWindowKeyMapping.cs
@@ -6,6 +6,7 @@ using System.Windows.Controls;
 using System.Windows.Input;
 using System.Windows.Media;
 
+using FalconBMS.Launcher.Core;
 using FalconBMS.Launcher.Windows;
 
 using Microsoft.DirectX.DirectInput;
