+++ FalconBMS Alternative Launcher Cs/packages.config
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="ControlzEx" version="3.0.2.4" targetFramework="net452" />
-  <package id="MahApps.Metro" version="1.6.5" targetFramework="net452" />
+  <package id="ControlzEx" version="3.0.2.4" targetFramework="net452" requireReinstallation="true" />
+  <package id="MahApps.Metro" version="1.6.5" targetFramework="net452" requireReinstallation="true" />
 </packages>
